+++ packages/contracts-communication/contracts/legacy/MessageBus.sol
@@ -93,7 +93,7 @@ contract MessageBus is ICAppV1, MessageBusEvents, IMessageBus {
     function _receiveMessage(
         uint64 srcChainId,
         bytes32, // sender
-        uint256, // dbNonce
+        uint64, // dbNonce
         uint64, // entryIndex
         bytes calldata encodedLegacyMsg
     )
